CLDR-17587 Update thousands separator for Haitian Creole to nbsp
I had an error (regular spaces shouldn't be used, only non-breaking spaces) and it caused me to check online media for the Haitian creole number formats. What I would write in English as "1,234,567.89" is shown in a few different ways, depending on the source.

Survey we at Meta gave to Haitian Creole linguists: "1 234 567,89"

[Google Translate](https://translate.google.com/?sl=en&tl=ht&text=1%2C234%2C567.89&op=translate): "1.234.567,89":

[Haitian Times](https://haitiantimes.com/2024/08/04/haitis-northwest-mythical-coast-anse-a-foleur-to-celebrate-140-years/): "1,234,567.89"

er, that's not great that the different sources work differently. I'll follow the survey'd results for initializing the core data but I'll follow up with Haitian Creole experts when we enter their language's data into the CLDR Survey Tool.